[Update] MainContainer Lookup Component postion Fix
