+++ src/Components/Main/Main.js
@@ -1,9 +1,26 @@
 import React from "react";
 import { MainWrapper } from "./Styled";
-import {Lookup} from "../index";
+import { Lookup } from "../index";
+import styled from "@emotion/styled";
 
+const MainLookupWrapper = styled.div`
+  position: absolute;
+  width: 100%;
+  min-width: 1440px;
+  height: 850px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 2;
+`;
 const Main = () => {
-  return <MainWrapper><Lookup/></MainWrapper>;
+  return (
+    <MainWrapper>
+      <MainLookupWrapper>
+        <Lookup />
+      </MainLookupWrapper>
+    </MainWrapper>
+  );
 };
 
 export default Main;
